black: preserve trailing commas
This doesn't yet cause the goal-state of exploding literals with
trailing commas, but it's close. Note that this is effectively a
reversion of 9394de150ebf0adc426523f46dc08e8b2b2b0b63, but once we
explode things across lines it'll be even more different.

Note that in already-blackened code, no formatting changes should occur.