black: tweak collection literals to explode with trailing comma
The change in behavior should be fairly clear from the test data change:
collection literals with trailing commas will be kept one-per-line
instead of being folded onto a single line if they fit. This matches
some logic in yapf and clang-format, and affords users a way to force
collections into a more merge-friendly format as desired.

Changes to existing tests (that is, outside collections.py) match my
expectations: the only new wrapping is on inputs that already contain
trailing commas, which was previously removed by black.

I regenerated expression.diff by using

    python3 -c 'print(open("tests/data/expression.py").read().split("# output")[0].strip())' | black - --diff --fast --config tests/data/empty_pyproject.toml > tests/data/expression.diff

and then reinstating the deterministic header. This saved me a fair
amount of time, but do be aware that the diff was machine-generated and
only loosely spot-checked since diff-of-diffs is hard even for this
source control hacker to read. :)