Stop trimming whitespace in docstrings
The PEP 257 algorithm used in #1053 results in trimming trailing
whitespace in docstrings -- see #1415.

Adjust the algorithm to preserve leading and trailing whitespace; this
reverts the changes in #1417.  This diverges from PEP 257, but better
retains the contents of the docstring.

Fixes #1452 because we no longer can end up with four trailing quotes.