Fix attribute error on black module in Vim
When using Black in Vim, the Vim plugin calls
a function on the black module called "find_pyproject_toml"
and "parse_pyproject_toml". However, these functions are not
yet part of the PyPI release of Black. Thus this Vim plugin
fails with an attribute error. This commit fixes that issue
as well as an issue where "skip_string_normalization" was not
properly being cast to a boolean value.