Refactor normalize_string_quotes
Keep behavior exactly the same, but refactor the internals of
normalize_string_quotes so that it is easier to grok the logic.

This also provides an (intentionally undocumented) hook for
organizations that are deeply resistant to the double-quotes
choice to avoid forking the black repository. A sufficiently clever
devops person could apply a single patch file to HEAD to select
the quote scheme that they desire. This is not tested, and is
therefore a "do at your own risk" procedure which is the maximal
extent to which it is appropriate for this project to support quote
preferences.