Switch prettier for mdformat (plus a few plugins)
prettier doesn't support MyST's syntax extensions which are going to be
used in this reorganization effort so we have to switch formatter.

Unfortanately mdformat's style is different from prettier's so time to
reformat the whole repo too.

We're excluding .github/ISSUE_TEMPLATE because I have no idea whether
its changes are safe, so let's play it safe.