Fix mypyc + Black on Windows
The `platform=linux` was causing mypy (and therefore mypyc) to believe
any `if sys.platform == "win32":` branch would never be taken. That led
to `Reached allegedly unreachable code!` crashes because of safety
checks mypyc adds.

There's not a strong reason for pinning the platform in `mypy.ini` so
with the agreement of Jelle, it's gone now!

I also disabled CliRunner's exception catching feature because while it
does lead to nicer test failures when an exception goes unhandled, it
also removes traceback information which makes debugging a pain.

P.S. You can blame Windows Update for the slowness of this bugfix :p