feat(config): Support `--config` in pyproject.toml
You can now specify `config` key in pyproject.toml which does the same
job as the CLI option `--config`. The path of config should be relative
to the CWD it is running from. If the file is not found or is invalid
i.e. it is a folder rather a TOML file it would raise
`click.exceptions.FileError`, and if the file can't be parsed it would
raise `tomli.TOMLDecodeError`. All black config keys in the custom
config should be registered under `[black]` and they would overwrite the
config specified in pyproject but not the one specified by `--config`
flag while running black with the CLI.

Closes #1826