feat(config): Add support for path specific config
Woohoo! Module specific configuration! and a bit of restructuring of code in the
`black/__init__.py` file. Mainly shifting all the "reformatting" logic to a
separate function to increase the readability of code in the click CLI
function. You can go through the code in the commit, code should be self
explanatory, though I still need to update the documentation of the
code and this feature.

I haven't aimed at bringing new tests in this commit, if required, I
could do that and this rather fixes the tests to work according to the
change I have done in this commit.

Summarising the changes:

- Now, you can support file/module specific configuration in the
  pyproject.toml config under the section of `[tools.black-module]`. I
  have added a separate "key" for the modules as according to the
  current code when the module name could clash with a config argument.
  For example `[tools.black.src]` would clash with `SRC`.
- Module specific configuration can only some arguments which are listed
  under a global constant in `config.py` and should agree with the
  type hint present on them in the class variables, the logic for this
  functionality can be seen in `cast_typehint` function of `config.py`.
  At some places I have ignored MYPY errors so as I couldn't change the
  configuration since they were being used for typing.
- Added a new class for storing all the parameters, the order of the
  arguments is as follows:
  * Read global pyproject.toml/specified config config.
  * Read module specific config, which would take the none specified
    arguments from the global configuration.
  * If value specified through click options/arguments it would be
    overwritten, this won't happen if the default value of the option is
    taken.
  * Tada!