Treat `--config` and `pyproject.toml` in the same way
There are two ways configs are loaded, (1) they are passed by the user through `--config` flag and (2) the `pyproject.toml` is found in the `src`. So currently when we find the `config` key we check whether if it is found by the config found in `pyproject.toml` or it is found in the one passed by the user via `--config` option.

In this commit, the above check has been dropped, and if config key found in any of the config, it would overwrite it no matter what.