Run second pass of formatter for real instead of in validation-only
Fixes #2488 and #2518

Admittedly, it doesn't fix the underlying issue. There are three
expected-failure tests in test_black.py, which prove the issue
(test_trailing_comma_optional_parens_stability{1,2,3})

This is a (somewhat terrible) bandaid which runs the entire formatter
twice to work around this problem. The problem appears preexisting
based on the disabled tests and the comments, so this feels like a
reasonable step-in-the-right-direction.